Basic Eating Implementation
- Changed the EntityHealth's health variable type to float
- Added PlayerHealth script and replaced the EntityHealth component on the player
- On the GrappleCombat, commented invincibility line and added increase player health

+++ Assets/Scripts/HUD/HUDCounterHandler.cs
@@ -6,7 +6,7 @@ public class HUDCounterHandler : MonoBehaviour
     [SerializeField] private int _imagePadding;
     [SerializeField] private int _imageTopPadding;
 
-    public void SetHUDCounter(int value)
+    public void SetHUDCounter(float value)
     {
 		for (int i = 0; i < transform.childCount; i++)
 		{
